Use util function for formatting
