+++ modules/polling/components/PollVoteHistoryItem.tsx
@@ -1,16 +1,16 @@
 /** @jsx jsx */
 import { getNetwork } from 'lib/maker';
-import { format } from 'date-fns';
 import Link from 'next/link';
 import { Box, Text, jsx, Link as ThemeUILink } from 'theme-ui';
 import { PollVoteHistory } from '../types/pollVoteHistory';
 import { PollVotePluralityResultsCompact } from './PollVotePluralityResultsCompact';
 import { Icon } from '@makerdao/dai-ui-icons';
+import { formatDateWithTime } from 'lib/datetime';
 import { POLL_VOTE_TYPE } from '../polling.constants';
 
 export function PollVoteHistoryItem({ vote }: { vote: PollVoteHistory }): React.ReactElement {
   const network = getNetwork();
-  const voteDate = format(new Date(vote.blockTimestamp), 'MMM dd yyyy HH:mm zz');
+  const voteDate = formatDateWithTime(vote.blockTimestamp);
   const isPluralityVote = vote.poll.voteType === POLL_VOTE_TYPE.PLURALITY_VOTE;
   const voteColorStyles = ['secondaryEmphasis', 'primary', 'notice'];
